requirements.txt
pillow
flask
gunicorn
requests
psycopg2-binary
